Simplify the message passing
We were passing a single struct across the
entire code. This obfuscated the fact that
we actually manipulated 2 different concepts:
1. The criteria for search and sorting games
(game name and target price)
2. The parsed game
(game name, id and price)

By splitting those 2, the logic becomes less
convoluted. We lose the ability to display
the target price though. This is not really
an issue as this is mostly used for debugging.