Improve error handling
We were swallowing errors in an goroutine
and never raising it to the main. This made
the code more complex as we had to comingle
management of the channel with the parsing.

The new code splits those 2 and always closes
the channel in the main, avoiding some side
effect that made it unclear when the channel
was valid.